server_deploy/requirements.txt
Flask
Gunicorn
gevent
pillow
opencv-python
numpy
Image


